requirements.txt
fastapi
uvicorn
google-generativeai
typing-extensions
openai
